# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API мини-приложения реверс-найма в MAX.\nАвторизация: POST /api/v1/auth c initData из MAX Bridge,\nдалее заголовок Authorization: Bearer \u003ctoken\u003e (JWT, 7 дней).",
        "title": "MAX Mini App — Reverse Hiring API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/invitations/{id}/respond": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitations"
                ],
                "summary": "Ответить на приглашение",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID приглашения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ответ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.RespondRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.MatchResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/my/invitations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitations"
                ],
                "summary": "Мои приглашения",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/max-miniapp_backend_internal_dto.Invitation"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/vacancies/{id}/candidates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "mode=list — постранично (limit/offset), mode=feed — лента без пагинации.\nТолько кандидаты с активным резюме и без действия по этой вакансии.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matching"
                ],
                "summary": "Кандидаты под вакансию",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID вакансии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "list",
                            "feed"
                        ],
                        "type": "string",
                        "description": "list",
                        "name": "mode",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "20",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.CandidatesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vacancies/{id}/candidates/{candidateUserId}/action": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matching"
                ],
                "summary": "Действие рекрутера на кандидата",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID вакансии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID кандидата",
                        "name": "candidateUserId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "action",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.RecruiterActionResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.RecruiterActionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "max-miniapp_backend_internal_dto.CandidateItem": {
            "type": "object",
            "properties": {
                "ai_comment": {
                    "type": "string"
                },
                "ai_score": {
                    "type": "integer"
                },
                "breakdown": {
                    "$ref": "#/definitions/max-miniapp_backend_internal_scoring.Breakdown"
                },
                "final_score": {
                    "type": "integer",
                    "example": 85
                },
                "resume": {
                    "$ref": "#/definitions/max-miniapp_backend_internal_dto.Resume"
                },
                "score": {
                    "type": "integer",
                    "example": 87
                },
                "user": {
                    "$ref": "#/definitions/max-miniapp_backend_internal_dto.User"
                }
            }
        },
        "max-miniapp_backend_internal_dto.CandidatesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/max-miniapp_backend_internal_dto.CandidateItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
        "max-miniapp_backend_internal_dto.Invitation": {
            "type": "object",
            "properties": {
                "company": {
                    "$ref": "#/definitions/max-miniapp_backend_internal_dto.InvitationCompany"
                },
                "deadline_at": {
                    "type": "string"
                },
                "hours_left": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "overdue",
                        "resulted"
                    ]
                },
                "vacancy": {
                    "$ref": "#/definitions/max-miniapp_backend_internal_dto.InvitationVacancy"
                }
            }
        },
        "max-miniapp_backend_internal_dto.InvitationCompany": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
        "max-miniapp_backend_internal_dto.InvitationVacancy": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.MatchResult": {
            "type": "object",
            "properties": {
                "company": {
                    "$ref": "#/definitions/max-miniapp_backend_internal_dto.InvitationCompany"
                },
                "id": {
                    "type": "integer"
                },
                "recruiter_contact": {
                    "type": "string"
                },
                "response": {
                    "type": "string"
                },
                "vacancy": {
                    "$ref": "#/definitions/max-miniapp_backend_internal_dto.InvitationVacancy"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.RecruiterActionResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "enum": [
                        "invite",
                        "skip"
                    ]
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.RespondRequest": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string",
                    "enum": [
                        "accept",
                        "decline"
                    ]
                }
            }
        },
        "max-miniapp_backend_internal_dto.Resume": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string",
                    "example": "Пишу сервисы на Go"
                },
                "city": {
                    "type": "string",
                    "example": "Москва"
                },
                "created_at": {
                    "type": "string"
                },
                "education": {
                    "type": "string",
                    "example": "МГТУ, ИТ-факультет"
                },
                "employment_type": {
                    "type": "string",
                    "enum": [
                        "full_time",
                        "part_time",
                        "contract",
                        "internship"
                    ],
                    "example": "full_time"
                },
                "experience_months": {
                    "type": "integer",
                    "example": 36
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_confirmed_at": {
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "salary_max": {
                    "type": "integer",
                    "example": 250000
                },
                "salary_min": {
                    "type": "integer",
                    "example": 150000
                },
                "skills": {
                    "type": "string",
                    "example": "go, postgres, docker"
                },
                "source": {
                    "type": "string",
                    "enum": [
                        "manual",
                        "file_parse"
                    ]
                },
                "source_text": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "example": "Backend-разработчик"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "work_format": {
                    "type": "string",
                    "enum": [
                        "onsite",
                        "hybrid",
                        "remote"
                    ],
                    "example": "remote"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "id": {
                    "type": "integer",
                    "example": 42
                },
                "language_code": {
                    "type": "string",
                    "example": "ru"
                },
                "last_name": {
                    "type": "string",
                    "example": "Petrov"
                },
                "photo_url": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "example": "candidate"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "example": "ivan"
                }
            }
        },
...
        "max-miniapp_backend_internal_scoring.Breakdown": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "integer"
                },
                "experience": {
                    "type": "integer"
                },
                "schedule": {
                    "type": "integer"
                },
                "skills": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Значение: \"Bearer \u003cjwt\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}